+++ bin/d2l-test-runner.js
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 import { argv } from 'node:process';
-import { runner } from '../src/server/cli/test.js';
+import { runner } from '../src/server/cli/test-runner.js';
 
 const options = await runner.getOptions(argv);
